blending-api/requirements.txt
numpy==1.18.2
Cython==0.29.17
Flask==1.1.2
torch==1.4.0
scipy==1.4.1
requests==2.23.0
gunicorn
opencv_python==4.2.0.34
Flask_Cors==3.0.8
torchvision==0.5.0
Pillow==7.1.2
